Remove repeater drop, delete old files
