+++ base_station/gui/src/components/LCMEcho.vue
@@ -82,8 +82,6 @@
           '/ra_control': false,
           '/ra_openloop_cmd': false,
           '/ra_pidconfig_cmd': false,
-          '/rr_drop_complete': false,
-          '/rr_drop_init': false,
           '/sa_control': false,
           '/sa_endeffector_cmd': false,
           '/sa_openloop_cmd': false,
@@ -127,8 +125,6 @@
           {'topic': '/pi_settings', 'type': 'PiSettings'},
           {'topic': '/radio_update', 'type': 'RadioSignalStrength' },
           {'topic': '/radio_setup', 'type': 'Signal'},
-          {'topic': '/rr_drop_complete', 'type': 'RepeaterDrop'},
-          {'topic': '/rr_drop_init', 'type': 'RepeaterDrop'},
           {'topic': '/sa_motors', 'type': 'SAMotors'},
           {'topic': '/sensors', 'type': 'Sensors'},
           {'topic': '/sensor_switch', 'type': 'SensorSwitch'},
