+++ base_station/gui/src/components/SATask.vue
@@ -59,7 +59,6 @@ import Cameras from './Cameras.vue'
 import RoverMap from './RoverMap.vue'
 import CommIndicator from './CommIndicator.vue'
 import OdometryReading from './OdometryReadingSA.vue'
-import WaypointEditor from './WaypointEditor.vue'
 import LCMBridge from 'lcm_bridge_client/dist/bridge.js'
 import SpectralData from './SpectralData.vue'
 import Chlorophyll from './Chlorophyll.vue'
@@ -265,7 +264,6 @@ export default {
     Cameras,
     CommIndicator,
     OdometryReading,
-    WaypointEditor,
     SpectralData,
     Chlorophyll,
     Ammonia,
