--- rover_msgs/RepeaterDrop.lcm
@@ -1,4 +1,0 @@
-package rover_msgs;
-
-struct RepeaterDrop {
-}
